requirements.txt
pyspark~=3.5.5
pyspark-stubs~=3.0.0.post3
numpy~=2.2.4
py4j~=0.10.9.7
PyMySQL~=1.1.1
tomlkit~=0.13.2
requests~=2.32.3
urllib3~=2.4.0
python-decouple~=3.8
jieba~=0.42.1
pillow~=11.2.1
Flask~=3.1.0
wordcloud~=1.9.4
pandas~=2.2.3
pyecharts~=2.0.8